big query update, continued
Adding in the 32 additional queries yesterday clued me in to the fact that bash scripts can't call variables within variable, which was the method I'd been using to declare the $DOMAIN variable within queries.txt. To accomodate for this, queries.txt is no longer. Now, cypher queries live in an array within cypheroth.sh.